requirements.txt
numpy==1.26.4
scipy==1.12.0
matplotlib==3.8.2
pandas==2.2.2
pyyaml==6.0.1
